finished modularization
