+++ index.html
@@ -2,10 +2,10 @@
 <head>
   <title>@tract</title>
   <link rel="stylesheet" type="text/css" href="/static/style.css">
-  <link rel="stylesheet" type="text/css" href="/static/tingle.min.css">
-  <script src="/static/tingle.min.js"></script>
-  <script src="/static/interact.min.js"></script>
-  <script src="/static/script.js"></script>
+  <link rel="stylesheet" type="text/css" href="/static/vendor/tingle.min.css">
+  <script src="/static/vendor/tingle.min.js"></script>
+  <script src="/static/vendor/interact.min.js"></script>
+  <script src="/static/attractio.js"></script>
   <link rel="icon" type="image/png" sizes="32x32" href="/static/favicon-32x32.png">
 </head>
 <body>
