+++ package.json
@@ -4,8 +4,10 @@
   "description": "collaborative poetry on le web!",
   "main": "script.js",
   "scripts": {
-    "test": "jest",
-    "start": "node server.js"
+    "test": "jest --verbose",
+    "start": "node server.js",
+    "dev": "concurrently \"mongod --dbpath=./data\" \"nodemon server.js\"",
+    "dev-fresh": "rm -r ./data && mkdir data && concurrently \"mongod --dbpath=./data\" \"nodemon server.js\""
   },
   "author": "",
   "license": "ISC",
